response-pain-point-extraction/requirements.txt
google-genai
numpy
scikit-learn
python-dotenv

